--- src/main/java/Duke.java
@@ -1,22 +1,0 @@
-import java.util.Scanner;
-public class Duke {
-    public static void main(String[] args) {
-//        String logo = " ____        _        \n"
-//                + "|  _ \\ _   _| | _____ \n"
-//                + "| | | | | | | |/ / _ \\\n"
-//                + "| |_| | |_| |   <  __/\n"
-//                + "|____/ \\__,_|_|\\_\\___|\n";
-//        System.out.println("Hello from\n" + logo);
-        System.out.println("Hello! I'm Duke\n What can I do for you?");
-        while (true){
-            Scanner scanner = new Scanner(System.in);
-            String inputString = scanner.nextLine();
-            if (inputString.equals("bye")) {
-                System.out.println("Bye. Hope to see you again soon!");
-                break;
-            } else {
-                System.out.println(inputString);
-            }
-        }
-    }
-}
